+++ src/borg/archiver/help_cmd.py
@@ -399,7 +399,7 @@ class HelpMixIn:
     helptext["placeholders"] = textwrap.dedent(
         """
         Repository URLs, ``--name``, ``-a`` / ``--match-archives``, ``--comment``
-        and ``--remote-path`` values support these placeholders:
+        and ``BORG_REMOTE_PATH`` values support these placeholders:
 
         {hostname}
             The (short) hostname of the machine.