+++ src/borg/testsuite/legacyrepository_test.py
@@ -877,7 +877,6 @@ def test_hints_behaviour(repository):
 
 def _get_mock_args():
     class MockArgs:
-        remote_path = "borg"
         umask = 0o077
         debug_topics = []
         rsh = None
@@ -967,7 +966,8 @@ def test_remote_ssh_cmd(remote_repository):
         assert remote_repository.ssh_cmd(Location("ssh://example.com/foo")) == ["ssh", "--foo", "example.com"]
 
 
-def test_remote_borg_cmd(remote_repository):
+def test_remote_borg_cmd(remote_repository, monkeypatch):
+    monkeypatch.delenv("BORG_REMOTE_PATH", raising=False)
     with remote_repository:
         assert remote_repository.borg_cmd(None, testing=True) == [sys.executable, "-m", "borg", "serve"]
         args = _get_mock_args()
@@ -975,7 +975,7 @@ def test_remote_borg_cmd(remote_repository):
         logging.getLogger().setLevel(logging.INFO)
         # note: test logger is on info log level, so --info gets added automagically
         assert remote_repository.borg_cmd(args, testing=False) == ["borg", "serve", "--info"]
-        args.remote_path = "borg-0.28.2"
+        monkeypatch.setenv("BORG_REMOTE_PATH", "borg-0.28.2")
         assert remote_repository.borg_cmd(args, testing=False) == ["borg-0.28.2", "serve", "--info"]
         args.debug_topics = ["something_client_side", "repository_compaction"]
         assert remote_repository.borg_cmd(args, testing=False) == [
@@ -985,6 +985,7 @@ def test_remote_borg_cmd(remote_repository):
             "--debug-topic=borg.debug.repository_compaction",
         ]
         args = _get_mock_args()
+        monkeypatch.delenv("BORG_REMOTE_PATH")
         assert remote_repository.borg_cmd(args, testing=False) == ["borg", "serve", "--info"]
         args.rsh = "ssh -i foo"
         remote_repository._args = args